Working Lambda/T Plots
The frequencyDisorder2D function makes a plot which is analogous to figure 1 of the sacramento paper. I coded it so that it would work with the updated data taking pipeline of the 16th onwards (using the scope trace analysis and updated cicero variables) and the earlier data (mainly from the 4th and the 9th). These latter sets look really bad though and don't make much sense. Maybe I did something wrong, but for now the 16th data is the only one that seems to make some sense when plotted